third_party/blink/web_tests/external/wpt/html/anonymous-iframe/worker-cookies.tentative.https.window_worker=service_worker-expected.txt

This is a testharness.js-based test.
PASS set cookies
PASS Worker spawned from normal iframe can access global cookies
FAIL Worker spawned from credentialless iframe can't access global cookies assert_equals: expected (undefined) undefined but got (string) "cookie_value"
Harness: the test ran to completion.
